classes

+++ maven.eclipse/src/main/java/dnd/bot/maven/eclipse/User/Character/Description/Personality/Personality.java
@@ -1,7 +1,7 @@
 package dnd.bot.maven.eclipse.User.Character.Description.Personality;
 
 public class Personality {
-    float Armor;
-    float Speed;
-    float PossessionBonus;
+    int Armor;
+    int Speed;
+    int PossessionBonus;
 }
